requirements.txt
evtx~=0.8.9
python-evtx~=0.7.4
PyYAML~=6.0.2